LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-right-container-margin-expected.txt

Testing the MacOSFullscreenMediaControls right container margin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


4 enabled buttons
PASS rightContainer.margin is 16

3 enabled buttons
PASS rightContainer.margin is 24

2 enabled buttons
PASS rightContainer.margin is 24

PASS successfullyParsed is true

TEST COMPLETE
